requirements.txt
# Core (used by smoke, config, storage)
duckdb>=0.9.0
PyYAML>=6.0
pandas>=2.0.0

# Trading calendar: US (NYSE) and HK (XHKG)
pandas_market_calendars>=4.0.0

# Ingestion: Yahoo prices and lake cache
yfinance>=0.2.0
pyarrow>=14.0.0

# Tests
pytest>=7.0.0

# Dashboard (D-2)
streamlit>=1.28.0
